angular/src/app/module/user/user.service.ts
import { Observable } from 'rxjs';

import { HttpClient } from '@angular/common/http';
import { Injectable, inject } from '@angular/core';
import { environment } from '@env/environment';

import { UserCreateto, User } from './user.model';

@Injectable({
    providedIn: 'root'
})
export class UserService {
    private readonly http = inject(HttpClient);

    getUsers(): Observable<User[]> {
        return this.http.get<User[]>(`${environment.apiUrl}/users`);
    }

    createUser(user: UserCreateto): Observable<User> {
        return this.http.post<User>(`${environment.apiUrl}/users`, user);
    }
}